+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+* Allow pure glob patterns without options in the config.
 * Add a TOML configuration option and deprecate JSON configurations.
 * Remove unnecessary re-sorting.